# generated/swagger.json
{
	"basePath": "/api",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"Count": {
			"description": "",
			"properties": {
				"count": {
					"type": "number",
					"format": "double",
					"description": ""
				}
			},
			"required": [
				"count"
			],
			"type": "object"
		},
		"Product": {
			"description": "",
			"properties": {
				"_id": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"info": {
					"type": "string",
					"description": ""
				},
				"price": {
					"type": "number",
					"format": "double",
					"description": ""
				}
			},
			"required": [
				"_id",
				"name",
				"price"
			],
			"type": "object"
		}
	},
	"info": {
		"description": "Present prices of the products",
		"license": {
			"name": "MIT"
		},
		"title": "productprices",
		"version": "1.0.0"
	},
	"paths": {
		"/v1/products/count": {
			"get": {
				"operationId": "ProductsCount",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Count"
						}
					}
				},
				"description": "Get count of the entities",
				"parameters": [
					{
						"description": "the value for filtering the results",
						"in": "query",
						"name": "search",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/v1/products/": {
			"get": {
				"operationId": "ProductsList",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/Product"
							}
						}
					}
				},
				"description": "Get list of the entities",
				"parameters": [
					{
						"description": "the value for filtering the results",
						"in": "query",
						"name": "search",
						"required": true,
						"type": "string"
					},
					{
						"description": "the value for ordering the results",
						"in": "query",
						"name": "sort",
						"required": true,
						"type": "string"
					},
					{
						"description": "the values results",
						"in": "query",
						"name": "limit",
						"required": true,
						"type": "number",
						"format": "double"
					},
					{
						"description": "for the limit",
						"in": "query",
						"name": "offset",
						"required": true,
						"type": "number",
						"format": "double"
					}
				]
			},
			"post": {
				"operationId": "ProductsCreate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Product"
						}
					}
				},
				"description": "Create new entity",
				"parameters": [
					{
						"description": "data to create",
						"in": "body",
						"name": "product",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Product"
						}
					}
				]
			},
			"delete": {
				"operationId": "ProductsClean",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Clean all the entries",
				"parameters": []
			}
		},
		"/v1/products/{id}": {
			"get": {
				"operationId": "ProductsGet",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Product"
						}
					}
				},
				"description": "Get entry details",
				"parameters": [
					{
						"description": "of the entity",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"put": {
				"operationId": "ProductsUpdate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Product"
						}
					}
				},
				"description": "Update entry details",
				"parameters": [
					{
						"description": "of the entity",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"description": "data to update",
						"in": "body",
						"name": "product",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Product"
						}
					}
				]
			},
			"delete": {
				"operationId": "ProductsDelete",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete entry",
				"parameters": [
					{
						"description": "of the entity",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/v1/prices/": {
			"get": {
				"operationId": "PricesParseUrl",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Parse all embedded url for the prices",
				"parameters": []
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {}
}